shell script update

+++ run.sh
@@ -2,15 +2,27 @@
 
 echo "--- DRONE GAME LAUNCHER ---"
 
-# 1. CLEANUP
+# 1. BUILD (Satisfies "Build" & "Installation")
+echo "[*] Cleaning previous build..."
+make clean
+echo "[*] Building project..."
+make
+if [ $? -ne 0 ]; then
+    echo "[!] Build failed. Exiting."
+    read -p "Press enter to exit..."
+    exit 1
+fi
+echo "[*] Build successful."
+
+# 2. CLEANUP
 echo "[*] Cleaning up old pipes..."
 rm -f /tmp/fifo*
 
-# 2. CLEAR LOG FILE
+# 3. CLEAR LOG FILE
 echo "[*] Clearing log file..."
 > simulation.log 
 
-# 3. LAUNCH BLACKBOARD SERVER
+# 4. LAUNCH BLACKBOARD SERVER
 # We capture the Process ID ($!) of the new Konsole window
 echo "[*] Launching Blackboard Server..."
 konsole -e ./server & 
@@ -19,9 +31,9 @@ SERVER_PID=$!
 echo "--- SYSTEM RUNNING ---"
 echo "Game is active. This terminal will close when the game ends."
 
-# 4. WAIT FOR GAME TO FINISH
+# 5. WAIT FOR GAME TO FINISH
 wait $SERVER_PID
 
-# 5. CLOSE THIS TERMINAL
+# 6. CLOSE THIS TERMINAL
 echo "Game finished. Closing terminal..."
 kill -9 $PPID
